Save a few instructions in ADC interrupt.
The standard mandates that bitwise operations promote
to a (16-bit) int. Casting to uint8 hints to the compiler
that it can use 8 bit operations.

+++ display_led.c
@@ -432,5 +432,5 @@ ISR(ADC_vect)
 {
     // Only care about top 3 bits, inverted
     uint8_t temp = ADCH;
-    display_brightness = (~temp) >> 5;
+    display_brightness = ((uint8_t)~temp) >> 5;
 }
